backend/ai-server/requirements.txt
pymysql==1.1.1
pandas==2.2.3
lifelines==0.30.0
scikit-learn==1.6.1
kneed==0.8.5
prophet==1.1.6
google-generativeai==0.8.4
python-dotenv==1.1.0
flask==3.1.0
flask-cors==5.0.1